Move netty channel intializer into its own named class.
